requirements.txt
Flask==2.3.2
gunicorn
google-genai
